refactor: split creating loaders
